Revert "Shard Transaction Hash v1 (#5560)"
This reverts commit f9cdf3a0bd678ea7234b0787286eea1a229a4eb0.
