+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/migration/BackfillTransactionHashMigration.java
@@ -23,45 +23,39 @@ package com.hedera.mirror.importer.migration;
 import com.google.common.base.Stopwatch;
 import java.io.IOException;
 import javax.inject.Named;
-
 import org.springframework.context.annotation.Lazy;
-import org.springframework.core.env.Environment;
-import org.springframework.core.env.Profiles;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import com.hedera.mirror.importer.MirrorProperties;
 import com.hedera.mirror.importer.config.Owner;
 import com.hedera.mirror.importer.parser.record.entity.EntityProperties;
 
-
 @Named
 public class BackfillTransactionHashMigration extends RepeatableMigration {
 
     private static final String BACKFILL_TRANSACTION_HASH_SQL = """
             begin;
-            truncate %1$s;
-            insert into %1$s (consensus_timestamp, hash, payer_account_id)
+            truncate transaction_hash;
+            insert into transaction_hash (consensus_timestamp, hash, payer_account_id)
             select consensus_timestamp, transaction_hash, payer_account_id
             from transaction
             where consensus_timestamp >= ?;
             commit;
             """;
+
     private static final String START_TIMESTAMP_KEY = "startTimestamp";
 
     private final EntityProperties entityProperties;
 
     private final JdbcTemplate jdbcTemplate;
-    private final boolean isV2;
 
     @Lazy
     public BackfillTransactionHashMigration(EntityProperties entityProperties,
                                             @Owner JdbcTemplate jdbcTemplate,
-                                            MirrorProperties mirrorProperties,
-                                            Environment environment) {
+                                            MirrorProperties mirrorProperties) {
         super(mirrorProperties.getMigration());
         this.entityProperties = entityProperties;
         this.jdbcTemplate = jdbcTemplate;
-        this.isV2 = environment.acceptsProfiles(Profiles.of("v2"));
     }
 
     @Override
@@ -78,9 +72,7 @@ public class BackfillTransactionHashMigration extends RepeatableMigration {
         }
 
         var stopwatch = Stopwatch.createStarted();
-
-        jdbcTemplate.update(String.format(BACKFILL_TRANSACTION_HASH_SQL, isV2 ? "transaction_hash" : "transaction_hash_sharded"), startTimestamp);
-
+        jdbcTemplate.update(BACKFILL_TRANSACTION_HASH_SQL, startTimestamp);
         log.info("Backfilled transaction hash for transactions at or after {} in {}", startTimestamp, stopwatch);
     }
