+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/batch/CompositeBatchPersister.java
@@ -20,17 +20,14 @@ package com.hedera.mirror.importer.parser.batch;
  * ‍
  */
 
-import com.hedera.mirror.common.domain.transaction.TransactionHash;
-
 import io.micrometer.core.instrument.MeterRegistry;
 import java.util.Collection;
 import java.util.Map;
-import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import javax.inject.Named;
 import javax.persistence.Entity;
 import javax.sql.DataSource;
-
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Primary;
 import org.springframework.core.annotation.AnnotationUtils;
 
@@ -41,6 +38,7 @@ import com.hedera.mirror.importer.repository.upsert.UpsertQueryGeneratorFactory;
 
 @Named
 @Primary
+@RequiredArgsConstructor
 public class CompositeBatchPersister implements BatchPersister {
 
     private final Map<Class<?>, BatchPersister> batchPersisters = new ConcurrentHashMap<>();
@@ -48,19 +46,6 @@ public class CompositeBatchPersister implements BatchPersister {
     private final MeterRegistry meterRegistry;
     private final CommonParserProperties properties;
     private final UpsertQueryGeneratorFactory upsertQueryGeneratorFactory;
-
-    public CompositeBatchPersister(DataSource dataSource, MeterRegistry meterRegistry,
-                                   CommonParserProperties properties,
-                                   UpsertQueryGeneratorFactory upsertQueryGeneratorFactory,
-                                   Optional<TransactionHashBatchInserter> transactionHashV1BatchPersister) {
-        this.dataSource = dataSource;
-        this.meterRegistry = meterRegistry;
-        this.properties = properties;
-        this.upsertQueryGeneratorFactory = upsertQueryGeneratorFactory;
-
-        transactionHashV1BatchPersister.ifPresent(batchPersister -> batchPersisters.put(TransactionHash.class,
-                batchPersister));
-    }
 
     @Override
     public void persist(Collection<? extends Object> items) {
