+++ hedera-mirror-importer/src/test/java/com/hedera/mirror/importer/TestUtils.java
@@ -22,8 +22,6 @@ package com.hedera.mirror.importer;
 
 import static java.lang.invoke.MethodType.methodType;
 
-import com.hedera.mirror.common.domain.transaction.TransactionHash;
-
 import com.hederahashgraph.api.proto.java.AccountID;
 import com.hederahashgraph.api.proto.java.Key;
 import com.hederahashgraph.api.proto.java.Timestamp;
@@ -48,9 +46,6 @@ import lombok.experimental.UtilityClass;
 import org.apache.commons.beanutils.BeanUtilsBean;
 
 import com.hedera.mirror.importer.util.Utility;
-
-import org.springframework.jdbc.core.DataClassRowMapper;
-import org.springframework.jdbc.core.JdbcTemplate;
 
 @UtilityClass
 public class TestUtils {
@@ -147,11 +142,6 @@ public class TestUtils {
         } catch (IllegalAccessException | InvocationTargetException e) {
             throw new RuntimeException(e);
         }
-    }
-
-    public Collection<TransactionHash> getShardTransactionHashes(int shard, JdbcTemplate jdbcTemplate) {
-        var sql = String.format("SELECT * from transaction_hash_sharded_%02d", shard);
-        return jdbcTemplate.query(sql, new DataClassRowMapper<>(TransactionHash.class));
     }
 
     public AccountID toAccountId(String accountId) {
