+++ hedera-mirror-importer/src/test/java/com/hedera/mirror/importer/repository/TransactionHashRepositoryTest.java
@@ -22,14 +22,11 @@ package com.hedera.mirror.importer.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.hedera.mirror.importer.EnabledIfV2;
-
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
-@EnabledIfV2
 class TransactionHashRepositoryTest extends AbstractRepositoryTest {
 
     private final TransactionHashRepository transactionHashRepository;
