+++ hedera-mirror-rest/__tests__/integration/transactionHashV1Matrix.js
@@ -20,11 +20,7 @@
 
 import {isV2Schema} from '../testutils.js';
 
-const nullifyPayerAccountId = async () => pool.queryQuietly('update transaction_hash_sharded set payer_account_id = null');
-const putHashInOldTable = async () => pool.queryQuietly(
-  `with deleted as (DELETE from transaction_hash_sharded RETURNING *)
-                    INSERT into transaction_hash_old(consensus_timestamp, hash, payer_account_id)
-                       SELECT consensus_timestamp, hash, payer_account_id from deleted`);
+const nullifyPayerAccountId = async () => pool.queryQuietly('update transaction_hash set payer_account_id = null');
 
 const applyMatrix = (spec) => {
   if (isV2Schema()) {
@@ -32,18 +28,13 @@ const applyMatrix = (spec) => {
   }
 
   const defaultSpec = {...spec};
-
   defaultSpec.name = `${defaultSpec.name} - default`;
 
   const nullPayerAccountIdSpec = {...spec};
   nullPayerAccountIdSpec.name = `${nullPayerAccountIdSpec.name} - null transaction_hash.payer_account_id`;
   nullPayerAccountIdSpec.postSetup = nullifyPayerAccountId;
 
-  const transactionHashOldSpec = {...spec};
-  transactionHashOldSpec.name = `${transactionHashOldSpec.name} - in old transaction_hash table`
-  transactionHashOldSpec.postSetup = putHashInOldTable;
-
-  return [defaultSpec, transactionHashOldSpec, nullPayerAccountIdSpec];
+  return [defaultSpec, nullPayerAccountIdSpec];
 };
 
 export default applyMatrix;
