+++ hedera-mirror-rest/__tests__/integrationDomainOps.js
@@ -26,7 +26,6 @@ import base32 from '../base32';
 import config from '../config';
 import * as constants from '../constants';
 import EntityId from '../entityId';
-import {isV2Schema, valueToBuffer} from "./testutils.js";
 
 const NETWORK_FEE = 1n;
 const NODE_FEE = 2n;
@@ -398,6 +397,27 @@ const loadTopicMessages = async (messages) => {
   for (const message of messages) {
     await addTopicMessage(message);
   }
+};
+
+const hexRegex = /^(0x)?[0-9A-Fa-f]+$/;
+
+const valueToBuffer = (value) => {
+  if (value === null) {
+    return value;
+  }
+
+  if (typeof value === 'string') {
+    if (hexRegex.test(value)) {
+      return Buffer.from(value.replace(/^0x/, '').padStart(2, '0'), 'hex');
+    }
+
+    // base64
+    return Buffer.from(value, 'base64');
+  } else if (Array.isArray(value)) {
+    return Buffer.from(value);
+  }
+
+  return value;
 };
 
 const convertByteaFields = (fields, object) => fields.forEach((field) => _.update(object, field, valueToBuffer));
@@ -824,11 +844,7 @@ const addTransaction = async (transaction) => {
 
 const addTransactionHash = async (transactionHash) => {
   transactionHash.hash = valueToBuffer(transactionHash.hash);
-  let table = 'transaction_hash';
-  if (!isV2Schema()) {
-    table = 'transaction_hash_sharded';
-  }
-  await insertDomainObject(table, Object.keys(transactionHash), transactionHash);
+  await insertDomainObject('transaction_hash', Object.keys(transactionHash), transactionHash);
 };
 
 const insertTransfers = async (
